+++ src/postorius/tests/test_list_forms.py
@@ -235,6 +235,8 @@ class TestListIdentityForm(TestCase):
             'info': 'This is a larger description of this mailing list.',
             'display_name': 'Most Desirable Mailing List',
             'subject_prefix': '  [Most Desirable]               ',
+            'preferred_language': 'en',
+            'member_roster_visibility': 'public',
         }, mlist=None)
         self.assertFalse(form.is_valid())
         self.assertTrue('advertised' in form.errors)